route dynamically set

+++ routes/api/admin.php
@@ -9,17 +9,17 @@ use Illuminate\Support\Facades\Route;
 Route::middleware(['auth:sanctum'])->group(function(){
     
     Route::middleware(['can:admin'])->prefix('admin')->group(function(){
-        Route::post('products', [ProductController::class ,'store'])->name('products.store');
-        Route::put('products/{slug}', [ProductController::class ,'update'])->name('products.update');
-        Route::delete('products/{slug}', [ProductController::class ,'destroy'])->name('products.destroy');
-        Route::get('products', [ProductController::class ,'index'])->name('products.index');
-        Route::get('products/{slug}', [ProductController::class ,'show'])->name('products.show');
+        Route::post('products', [ProductController::class ,'store'])->name('admin.products.store');
+        Route::put('products/{slug}', [ProductController::class ,'update'])->name('admin.products.update');
+        Route::delete('products/{slug}', [ProductController::class ,'destroy'])->name('admin.products.destroy');
+        Route::get('products', [ProductController::class ,'index'])->name('admin.products.index');
+        Route::get('products/{slug}', [ProductController::class ,'show'])->name('admin.products.show');
         
         
-        Route::post('categories', [CategoryController::class ,'store'])->name('categories.store');
-        Route::put('categories/{slug}', [CategoryController::class ,'update'])->name('categories.update');
-        Route::delete('categories/{slug}', [CategoryController::class ,'destroy'])->name('categories.destroy');
-        Route::get('categories', [CategoryController::class ,'index'])->name('categories.index');
-        Route::get('categories/{slug}', [CategoryController::class ,'show'])->name('categories.show');
+        Route::post('categories', [CategoryController::class ,'store'])->name('admin.categories.store');
+        Route::put('categories/{slug}', [CategoryController::class ,'update'])->name('admin.categories.update');
+        Route::delete('categories/{slug}', [CategoryController::class ,'destroy'])->name('admin.categories.destroy');
+        Route::get('categories', [CategoryController::class ,'index'])->name('admin.categories.index');
+        Route::get('categories/{slug}', [CategoryController::class ,'show'])->name('admin.categories.show');
     });
 });
